requirements.txt
streamlit
python-dotenv
openai>=1.35.7
openai-agents
